+++ src/themes/_generator/tokens/fontFamily/fontFamily.types.ts
@@ -1,2 +1,2 @@
-export type Name = "title" | "body";
+export type Name = "title" | "body" | "monospace";
 export type Token = { family: string };